Refactor show command directory structure
